fix(review): apply the remaining structural and test findings
--reuse-db in the shared addopts kept an existing test database on the old
schema, so a developer who did not pass --create-db met a confusing failure
after a new migration. The default is now correct, and the devcontainer
wrapper opts back in for its fast local loop; a caller's --create-db still
wins because the wrapper puts its own flags first. CI installs from
requirements_dev.txt so it picks up the floors that file declares.

Production changes:

- Move _add_vlan_group_selection and _add_missing_vlans_info into
  VlanAssignmentMixin. The verify view called them unbound off
  BaseInterfaceTableView while passing its own instance, which worked only
  because neither method touched a member the mixin does not provide. Any
  later self access there would have broken silently.
- Build the relationship port-name list once and pass it to the three signal
  checks instead of recomputing it per check.
- Drop the register_model_view decorators from the mapping bulk-import views.
  urls.py never includes get_model_urls(), so they registered no URL; the
  explicit routes already carry these views and a new test pins them.
- Give the new PortStackLagPattern serializer url, display and brief_fields,
  so brief and nested representations match the NetBox 4.x contract. The
  sibling serializers predate this stack and are left alone.
- Prefix the interface-name radio element IDs. They were page-global IDs
  emitted from a shared include, so rendering it twice broke the label
  association and the radio group.

Test changes:

- Give the VLAN-scope race test a positive control. The old assertion held
  whenever no VLAN was assigned at all, so a wrong POST key or a skipped row
  passed it. The same flow now runs once without the race and asserts VLAN
  100 is assigned, which is what makes the empty result meaningful.
- Use SET LOCAL for the module-adoption lock timeout. A session SET survives
  the test transaction and leaks into every later test on that connection.
- Clear the cache namespace on teardown as well, so a long parallel run does
  not accumulate dead keys until their TTL expires.
- Pass type_label to the badge-contrast helper, so it inspects the markup
  render_parent actually produces.
- Name the blocked-thread wait window once and allow the environment to raise
  it, instead of four separate literals.
- Document the private ContentType cache coupling and name the supported
  alternative.
- Drop the worker count from the alias help text, where it could drift from
  MAX_PARALLEL_WORKERS.
