+++ netbox_librenms_plugin/views/mapping_views.py
@@ -1,7 +1,6 @@
 from django.http import HttpResponse, HttpResponseBadRequest
 from django.views import View
 from netbox.views import generic
-from utilities.views import register_model_view
 
 from netbox_librenms_plugin.filters import (
     CarrierAutoInstallRuleFilterSet,
@@ -93,7 +92,6 @@ class InterfaceTypeMappingCreateView(LibreNMSWritePermissionMixin, generic.Objec
     form = InterfaceTypeMappingForm
 
 
-@register_model_view(InterfaceTypeMapping, "bulk_import", path="import", detail=False)
 class InterfaceTypeMappingBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """
     Provides a view for bulk importing `InterfaceTypeMapping` objects from CSV, JSON, or YAML.
@@ -166,7 +164,6 @@ class DeviceTypeMappingCreateView(LibreNMSWritePermissionMixin, generic.ObjectEd
     form = DeviceTypeMappingForm
 
 
-@register_model_view(DeviceTypeMapping, "bulk_import", path="import", detail=False)
 class DeviceTypeMappingBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing DeviceTypeMapping objects."""
 
@@ -226,7 +223,6 @@ class ModuleTypeMappingCreateView(LibreNMSWritePermissionMixin, generic.ObjectEd
     form = ModuleTypeMappingForm
 
 
-@register_model_view(ModuleTypeMapping, "bulk_import", path="import", detail=False)
 class ModuleTypeMappingBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing ModuleTypeMapping objects."""
 
@@ -286,7 +282,6 @@ class ModuleBayMappingCreateView(LibreNMSWritePermissionMixin, generic.ObjectEdi
     form = ModuleBayMappingForm
 
 
-@register_model_view(ModuleBayMapping, "bulk_import", path="import", detail=False)
 class ModuleBayMappingBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing ModuleBayMapping objects."""
 
@@ -346,7 +341,6 @@ class NormalizationRuleCreateView(LibreNMSWritePermissionMixin, generic.ObjectEd
     form = NormalizationRuleForm
 
 
-@register_model_view(NormalizationRule, "bulk_import", path="import", detail=False)
 class NormalizationRuleBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing NormalizationRule objects."""
 
@@ -406,7 +400,6 @@ class InventoryIgnoreRuleCreateView(LibreNMSWritePermissionMixin, generic.Object
     form = InventoryIgnoreRuleForm
 
 
-@register_model_view(InventoryIgnoreRule, "bulk_import", path="import", detail=False)
 class InventoryIgnoreRuleBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing InventoryIgnoreRule objects."""
 
@@ -526,7 +519,6 @@ class PlatformMappingCreateView(LibreNMSWritePermissionMixin, generic.ObjectEdit
     form = PlatformMappingForm
 
 
-@register_model_view(PlatformMapping, "bulk_import", path="import", detail=False)
 class PlatformMappingBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing PlatformMapping objects."""
 
@@ -586,7 +578,6 @@ class CarrierAutoInstallRuleCreateView(LibreNMSWritePermissionMixin, generic.Obj
     form = CarrierAutoInstallRuleForm
 
 
-@register_model_view(CarrierAutoInstallRule, "bulk_import", path="import", detail=False)
 class CarrierAutoInstallRuleBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing CarrierAutoInstallRule objects."""
 
@@ -650,7 +641,6 @@ class PortStackLagPatternCreateView(LibreNMSWritePermissionMixin, generic.Object
     form = PortStackLagPatternForm
 
 
-@register_model_view(PortStackLagPattern, "bulk_import", path="import", detail=False)
 class PortStackLagPatternBulkImportView(LibreNMSWritePermissionMixin, generic.BulkImportView):
     """Provides a view for bulk importing PortStackLagPattern objects from CSV/JSON/YAML."""
